server/requirements.txt
# CookClip Backend - Production Dependencies
# Core web framework
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.30.0,<0.32.0

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0

# Configuration
python-dotenv>=1.0.0,<2.0.0

# Vertex AI / Google Cloud
google-cloud-aiplatform>=1.38.0,<2.0.0

# YouTube & Audio
youtube-transcript-api>=0.6.2,<1.0.0
yt-dlp>=2025.1.0,<2026.0.0

# Google Cloud Speech-to-Text
google-cloud-speech>=2.21.0,<3.0.0

# Data validation
pydantic[email]>=2.9.0,<3.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0
google-auth>=2.23.0,<3.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
google-auth-httplib2>=0.2.0,<1.0.0

